refactor: User class 분리
