+++ src/libs/constant.js
@@ -4,3 +4,5 @@ export const USER_INPUT_ALERT = {
   removeZero: "1~9까지의 숫자만 입력해주세요",
   overlap: "중복되지 않은 숫자로 입력해주세요.",
 };
+
+export const NUMBER_LENGTH = 3;
